requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.20.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
tqdm>=4.61.0 